requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.3
django-ckeditor==5.3.0
django-cors-headers==2.1.0
django-js-asset==0.1.1
django-mysql==2.1.0
djangorestframework==3.6.3
idna==2.5
mysqlclient==1.3.10
psycopg2==2.7.1
pytz==2017.2
requests==2.18.1
simplejson==3.11.1
urllib3==1.21.1